lekker een keertje committen
